+++ mipidsi/src/models/ili9341.rs
@@ -40,10 +40,6 @@ impl Model for ILI9341Rgb565 {
     {
         ili934x::write_pixels_rgb565(dcs, colors)
     }
-
-    fn repeat_pixel_to_buffer(color: Self::ColorFormat, buf: &mut [u8]) -> Result<usize, Error> {
-        crate::graphics::repeat_pixel_to_buffer_rgb565(color, buf)
-    }
 }
 
 impl Model for ILI9341Rgb666 {
@@ -70,9 +66,5 @@ impl Model for ILI9341Rgb666 {
         I: IntoIterator<Item = Self::ColorFormat>,
     {
         ili934x::write_pixels_rgb666(dcs, colors)
-    }
-
-    fn repeat_pixel_to_buffer(color: Self::ColorFormat, buf: &mut [u8]) -> Result<usize, Error> {
-        crate::graphics::repeat_pixel_to_buffer_rgb666(color, buf)
     }
 }